feat: make shifts work again with 64 bit

+++ jolt/jolt-core/src/jolt/subtable/and.rs
@@ -19,7 +19,7 @@ impl<F: JoltField> AndSubtable<F> {
 }
 
 impl<F: JoltField> LassoSubtable<F> for AndSubtable<F> {
-    fn materialize(&self, M: usize) -> Vec<u32> {
+    fn materialize(&self, M: usize) -> Vec<u64> {
         // table[x | y] = x & y
         let mut entries = Vec::with_capacity(M);
         let bits_per_operand = (log2(M) / 2) as usize;
@@ -27,7 +27,7 @@ impl<F: JoltField> LassoSubtable<F> for AndSubtable<F> {
         // Materialize table entries in order where (x | y) ranges 0..M
         for idx in 0..M {
             let (x, y) = split_bits(idx, bits_per_operand);
-            let row = (x & y) as u32;
+            let row = (x & y) as u64;
             entries.push(row);
         }
         entries
